+++ README.md
@@ -1,4 +1,4 @@
-# Clockify Timer
+# Moondial Timer
 A tiny stopwatch, chronometer, and countdown timer library with millisecond precision built in TypeScript. Support for state change events implemented via RxJs. Use it in the browser or NodeJs.
 
 See the docs for live examples of the library in action: http://clockify.siliconcosmos.com
