+++ index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8">
 
-    <title>clockify timer</title>
+    <title>moondial timer</title>
 
     <link rel="stylesheet" href="src/main/demo/skeleton-main.css">
     <link rel="stylesheet" href="src/main/demo/skeleton-dark.css">
@@ -29,7 +29,7 @@
 
         <div>
             <a class="button button-primary" style="float: right;" href="./reference/">API Reference</a>
-            <h1>// clockify timer</h1>
+            <h1>// moondial timer</h1>
         </div>
 
         <p>
